+++ src/components/button-base/button-base.module.css.d.ts
@@ -6,12 +6,13 @@
  * actual or intended publication of such source code.
  */
 export const root: string;
-export const icon: string;
-export const singleIcon: string;
+export const active: string;
+export const hovered: string;
+export const focused: string;
+export const disabled: string;
+export const normal: string;
 export const ghost: string;
+export const intentDefault: string;
 export const accent: string;
-export const nowrap: string;
-export const small: string;
-export const tiny: string;
-export const rounded: string;
-export const text: string;
+export const alarm: string;
+export const inverted: string;